initial commit with styles
