--- app/src/main/java/com/peculiaruc/alc_mmsystem_admin/MainActivity.kt
@@ -1,11 +1,0 @@
-package com.peculiaruc.alc_mmsystem_admin
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
