--- app/src/main/java/com/peculiaruc/alc_mmsystem_admin/NewPasswordFragment.kt
@@ -1,27 +1,0 @@
-package com.peculiaruc.alc_mmsystem_admin
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-
-class NewPasswordFragment : Fragment(R.layout.fragment_new_password) {
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_new_password, container, false)
-    }
-
-
-}
