+++ app/src/main/java/com/peculiaruc/alc_mmsystem_admin/ui/activities/SplashActivity.kt
@@ -1,11 +1,11 @@
-package com.peculiaruc.alc_mmsystem_admin
+package com.peculiaruc.alc_mmsystem_admin.ui.activities
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
-import android.view.WindowManager
+import com.peculiaruc.alc_mmsystem_admin.R
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -13,7 +13,7 @@ class SplashActivity : AppCompatActivity() {
         setContentView(R.layout.activity_splash)
 
         Handler(Looper.getMainLooper()).postDelayed({
-            val intent = Intent(this@SplashActivity, LoginActivity::class.java)
+            val intent = Intent(this@SplashActivity, MainActivity::class.java)
             startActivity(intent)
             finish()
         }, 3000)
