+++ app/src/main/java/com/peculiaruc/alc_mmsystem_admin/ui/fragments/NewPasswordFragment.kt
@@ -1,0 +1,31 @@
+package com.peculiaruc.alc_mmsystem_admin.ui.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.peculiaruc.alc_mmsystem_admin.R
+import com.peculiaruc.alc_mmsystem_admin.databinding.FragmentNewPasswordBinding
+
+
+class NewPasswordFragment : Fragment() {
+
+    private lateinit var binding: FragmentNewPasswordBinding
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        super.onCreateView(inflater, container, savedInstanceState)
+
+        binding = FragmentNewPasswordBinding.inflate(layoutInflater)
+
+        binding.apply {
+
+        }
+
+        return binding.root
+    }
+
+}
