+++ app/src/main/res/layout/fragment_forgot_password.xml
@@ -5,8 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ForgotPasswordMessageFragment">
-
+    tools:context=".ui.fragments.ForgotPasswordFragment">
 
     <TextView
         android:id="@+id/forgotPasswordTitle"
@@ -37,7 +36,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/forgotPasswordTitle" />
 
-    <Button
+    <com.google.android.material.button.MaterialButton
         android:id="@+id/button_done"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -46,6 +45,8 @@
         android:fontFamily="@font/mukta_semi_bold"
         android:textSize="18sp"
         android:text="@string/done"
+        app:cornerRadius="10dp"
+        android:paddingVertical="15dp"
         app:layout_constraintEnd_toEndOf="@+id/forgotPasswordMessage"
         app:layout_constraintStart_toStartOf="@+id/forgotPasswordMessage"
         app:layout_constraintTop_toBottomOf="@+id/forgotPasswordMessage" />
