+++ app/src/main/res/layout/fragment_login.xml
@@ -5,29 +5,37 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".LoginActivity">
+    tools:context=".ui.fragments.LoginFragment">
 
-    <EditText
+    <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/editTextPassword"
+        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="32dp"
         android:layout_marginEnd="24dp"
-        android:ems="10"
+        app:passwordToggleEnabled="true"
+        app:endIconMode="password_toggle"
+        app:endIconTint="@color/black2"
+        app:passwordToggleTintMode="screen"
         android:fontFamily="@font/mukta_regular"
         android:hint="@string/password"
-        android:inputType="textPassword"
-        android:padding="16dp"
-        android:textAlignment="textStart"
         android:textColor="@color/black2"
         android:textColorHint="@color/black5"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="@+id/textView2"
-        app:layout_constraintTop_toBottomOf="@+id/editTextTextEmailAddress" />
+        app:layout_constraintStart_toStartOf="@+id/textView_sign_in_to_continue"
+        app:boxStrokeColor="@color/black3"
+        app:layout_constraintTop_toBottomOf="@+id/editTextTextEmailAddress">
+        <com.google.android.material.textfield.TextInputEditText
+            android:layout_width="match_parent"
+            android:textColor="@color/black2"
+            android:layout_height="wrap_content"
+            android:inputType="textPassword"/>
+    </com.google.android.material.textfield.TextInputLayout>
+
 
     <TextView
-        android:id="@+id/textView"
+        android:id="@+id/textView_welcome"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
@@ -41,43 +49,50 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/textView2"
+        android:id="@+id/textView_sign_in_to_continue"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:fontFamily="@font/mukta_regular"
         android:text="@string/sign_in_to_continue"
         android:textColor="@color/black2"
         android:textSize="16sp"
-        app:layout_constraintStart_toStartOf="@+id/textView"
-        app:layout_constraintTop_toBottomOf="@+id/textView" />
+        app:layout_constraintStart_toStartOf="@+id/textView_welcome"
+        app:layout_constraintTop_toBottomOf="@+id/textView_welcome" />
 
-    <EditText
+    <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/editTextTextEmailAddress"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="24dp"
         android:layout_marginEnd="24dp"
         android:ems="10"
+        app:boxStrokeColor="@color/black3"
+        android:textColorHint="@color/black5"
+        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
         android:fontFamily="@font/mukta_regular"
         android:hint="@string/email"
-        android:inputType="textEmailAddress"
-        android:padding="16dp"
-        android:textAlignment="textStart"
-        android:textColor="@color/black2"
-        android:textColorHint="@color/black5"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="@+id/textView2"
-        app:layout_constraintTop_toBottomOf="@+id/textView2" />
+        app:layout_constraintStart_toStartOf="@+id/textView_sign_in_to_continue"
+        app:layout_constraintTop_toBottomOf="@+id/textView_sign_in_to_continue">
+        <com.google.android.material.textfield.TextInputEditText
+            android:layout_width="match_parent"
+            android:textColor="@color/black2"
+            android:inputType="textEmailAddress"
+            android:layout_height="wrap_content"/>
+    </com.google.android.material.textfield.TextInputLayout>
 
-    <Button
+
+    <com.google.android.material.button.MaterialButton
         android:id="@+id/buttonLogin"
         android:layout_width="0dp"
-        android:layout_height="56dp"
+        android:layout_height="wrap_content"
+        android:paddingVertical="15dp"
         android:layout_marginStart="24dp"
         android:layout_marginTop="64dp"
         android:layout_marginEnd="24dp"
+        app:cornerRadius="10dp"
         android:fontFamily="@font/mukta_semi_bold"
-        android:text="Login"
+        android:text="@string/login"
         android:textAllCaps="false"
         android:textSize="18sp"
         app:layout_constraintEnd_toEndOf="parent"
@@ -106,21 +121,26 @@
         app:layout_constraintEnd_toEndOf="@+id/buttonLogin"
         app:layout_constraintTop_toBottomOf="@+id/buttonLogin" />
 
-    <Button
+    <com.google.android.material.button.MaterialButton
         android:id="@+id/button_signin_google"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="96dp"
-
+        android:outlineProvider="none"
+        android:paddingVertical="15dp"
+        android:textAppearance="@style/TextAppearance.AppCompat.Body1"
         android:backgroundTint="@color/white"
+        app:cornerRadius="10sp"
+        app:strokeColor="@color/pry2"
+        app:strokeWidth="1dp"
         android:drawableLeft="@drawable/google_logo"
-        android:fontFamily="@font/mukta_medium"
+        android:fontFamily="@font/mukta_semi_bold"
         android:gravity="center"
-        android:text="Sign  in with Google"
+        android:text="Sign in with Google"
         android:textAlignment="inherit"
         android:textAllCaps="false"
-        android:textColor="@color/pry1"
-        android:textSize="16sp"
+        android:textColor="@color/pry2"
+        android:textSize="18sp"
         app:iconGravity="textStart"
         app:layout_constraintEnd_toEndOf="@+id/forgot_password_text"
         app:layout_constraintHorizontal_bias="0.0"
@@ -128,7 +148,7 @@
         app:layout_constraintTop_toBottomOf="@+id/forgot_password_text" />
 
     <TextView
-        android:id="@+id/textView4"
+        android:id="@+id/textView_new_user"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="32dp"
@@ -144,7 +164,7 @@
         android:layout_marginStart="8dp"
         android:text="Sign up"
         android:textColor="@color/pry2"
-        app:layout_constraintStart_toEndOf="@+id/textView4"
-        app:layout_constraintTop_toTopOf="@+id/textView4" />
+        app:layout_constraintStart_toEndOf="@+id/textView_new_user"
+        app:layout_constraintTop_toTopOf="@+id/textView_new_user" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
