+++ app/src/main/res/layout/item_program.xml
@@ -1,0 +1,67 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:background="@color/white"
+    android:paddingVertical="10dp"
+    android:paddingHorizontal="10dp">
+
+    <com.google.android.material.imageview.ShapeableImageView
+        android:id="@+id/imageView_program"
+        android:layout_width="56dp"
+        android:layout_height="56dp"
+        app:shapeAppearanceOverlay="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="16dp"
+        android:cropToPadding="false"
+        android:scaleType="centerCrop"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:srcCompat="@tools:sample/backgrounds/scenic" />
+
+    <TextView
+        android:id="@+id/textView_message"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="16dp"
+        android:fontFamily="@font/mukta_semi_bold"
+        android:text="Peculiah C. Umeh"
+        android:textColor="@color/black2"
+        android:textSize="16sp"
+        app:layout_constraintBottom_toTopOf="@+id/textView_message_snap"
+        app:layout_constraintEnd_toStartOf="@+id/textView_time"
+
+        app:layout_constraintStart_toEndOf="@+id/imageView_program"
+        app:layout_constraintTop_toTopOf="@+id/imageView_program" />
+
+    <TextView
+        android:id="@+id/textView_message_snap"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="4dp"
+        android:textSize="12sp"
+        android:text="Can we go ahead to join the UI/UX team"
+        app:layout_constraintEnd_toEndOf="@+id/textView_message"
+        app:layout_constraintStart_toStartOf="@+id/textView_message"
+        app:layout_constraintTop_toBottomOf="@+id/textView_message"
+        android:ellipsize="end"
+        android:maxLines="1"
+        android:textColor="@color/black5"
+        android:fontFamily="@font/mukta_regular"
+        />
+
+    <TextView
+        android:id="@+id/textView_time"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:fontFamily="@font/mukta_regular"
+        android:text="30m"
+        android:textSize="12sp"
+        android:textColor="@color/black5"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/textView_message" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
