--- app/src/main/res/values-night/themes.xml
@@ -1,16 +1,0 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
-    <!-- Base application theme. -->
-    <style name="Theme.AlcMMSystemadmin" parent="Theme.MaterialComponents.NoActionBar">
-        <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/pry2</item>
-        <item name="colorPrimaryVariant">@color/pry1</item>
-        <item name="colorOnPrimary">@color/black2</item>
-        <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/pry4</item>
-        <item name="colorSecondaryVariant">@color/pry3</item>
-        <item name="colorOnSecondary">@color/black2</item>
-        <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
-        <!-- Customize your theme here. -->
-    </style>
-</resources>
