+++ app/src/main/res/values/themes.xml
@@ -6,11 +6,12 @@
         <item name="colorPrimaryVariant">@color/pry1</item>
         <item name="colorOnPrimary">@color/white</item>
         <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/pry4</item>
-        <item name="colorSecondaryVariant">@color/pry3</item>
+        <item name="colorSecondary">@color/pry2</item>
+        <item name="colorSecondaryVariant">@color/pry1</item>
         <item name="colorOnSecondary">@color/black2</item>
         <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
+        <item name="android:statusBarColor" >@color/white</item>
+        <item name="android:windowLightStatusBar" tools:targetApi="m">true</item>
         <!-- Customize your theme here. -->
     </style>
 </resources>
